+++ pgo/pgo.c
@@ -100,12 +100,12 @@ void wayl_win_alpha_changed(struct wl_window *win) {}
 bool wayl_win_set_urgent(struct wl_window *win) { return true; }
 bool wayl_fractional_scaling(const struct wayland *wayl) { return true; }
 
-bool
+pid_t
 spawn(struct reaper *reaper, const char *cwd, char *const argv[],
       int stdin_fd, int stdout_fd, int stderr_fd,
-      const char *xdg_activation_token)
+      reaper_cb cb, void *cb_data, const char *xdg_activation_token)
 {
-    return true;
+    return 2;
 }
 
 pid_t
@@ -151,8 +151,31 @@ void ime_enable(struct seat *seat) {}
 void ime_disable(struct seat *seat) {}
 void ime_reset_preedit(struct seat *seat) {}
 
+bool
+notify_notify(struct terminal *term, struct notification *notif)
+{
+    return true;
+}
+
 void
-notify_notify(const struct terminal *term, const char *title, const char *body)
+notify_free(struct terminal *term, struct notification *notif)
+{
+}
+
+void
+notify_icon_add(struct terminal *term, const char *id,
+                const char *symbolic_name, const uint8_t *data,
+                size_t data_sz)
+{
+}
+
+void
+notify_icon_del(struct terminal *term, const char *id)
+{
+}
+
+void
+notify_icon_free(struct notification_icon *icon)
 {
 }